report-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
python-dateutil
boto3
reportlab